SSL verifier fix
I have added the method to specify the certificate file. ie (PEM). I
have added a working example to the test folder along with the
certificate. You can verify it and also you can download the ssl
certifcate from firefox for any website that offers ssl and when saving
save it as .pem file instead of .crt and choose the option to download
with chain.